backend/pyproject.toml
[project]
name = "instalily-backend"
version = "0.1.0"
description = "FastAPI backend for PartSelect AI chat agent"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.2.0",
    "google-genai>=1.0.0",
    "chromadb>=0.6.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.2.2",
]
